blog_backend/requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
uvicorn 
gunicorn